+++ state-engine/src/main/java/content/LockContent.java
@@ -30,7 +30,7 @@ public abstract class LockContent implements Content {
     }
 
     @Override
-    public long GetLWM() {
+    public double GetLWM() {
         throw new UnsupportedOperationException();
     }
 
@@ -70,7 +70,7 @@ public abstract class LockContent implements Content {
     }
 
     @Override
-    public void AddLWM(long ts) {
+    public void AddLWM(double ts) {
         throw new UnsupportedOperationException();
     }
