+++ state-engine/src/main/java/scheduler/impl/op/structured/OPDFSAScheduler.java
@@ -58,7 +58,7 @@ public class OPDFSAScheduler<Context extends OPSAContext> extends OPDFSScheduler
     }
 
     @Override
-    public void PROCESS(Context context, long mark_ID) {
+    public void PROCESS(Context context, double mark_ID) {
         int cnt = 0;
         int batch_size = 100;//TODO;
         int threadId = context.thisThreadId;
@@ -143,7 +143,7 @@ public class OPDFSAScheduler<Context extends OPSAContext> extends OPDFSScheduler
      * @return
      */
     private boolean _MarkOperationsToAbort(Context context, Operation operation) {
-        long bid = operation.bid;
+        double bid = operation.bid;
         boolean markAny = false;
         //identify bids to be aborted.
         for (Operation failedOp : failedOperations) {
