+++ state-engine/src/main/java/scheduler/struct/og/Operation.java
@@ -43,7 +43,7 @@ public class Operation extends AbstractOperation implements Comparable<Operation
 
 
     public <Context extends OGSchedulerContext> Operation(String pKey, Function function, String table_name, SchemaRecordRef record_ref, TableRecord[] condition_records, Condition condition, int[] success,
-                     TxnContext txn_context, CommonMetaTypes.AccessType accessType, String operator_name, TableRecord s_record, TableRecord d_record, long bid, Context context) {
+                     TxnContext txn_context, CommonMetaTypes.AccessType accessType, String operator_name, TableRecord s_record, TableRecord d_record, double bid, Context context) {
         super(function, table_name, record_ref, condition_records, condition, success, txn_context, accessType, operator_name, s_record, d_record, bid);
 
         this.pKey = pKey;
@@ -79,7 +79,7 @@ public class Operation extends AbstractOperation implements Comparable<Operation
             }
             return this.d_record.getID() - operation.d_record.getID();
         } else
-            return Long.compare(this.bid, operation.bid);
+            return Double.compare(this.bid, operation.bid);
     }
 
 //    public void setConditionSources(String[] condition_sourceTable, String[] condition_source) {
