+++ state-engine/src/main/java/transaction/context/TxnContext.java
@@ -5,12 +5,12 @@ import storage.SchemaRecordRef;
 public class TxnContext {
     public final int thread_Id;
     private final int fid;
-    private final long bid;
+    private final double bid;
     private final String thisOpId;
     private final boolean is_read_only_;
     private final boolean is_dependent_;
     private final boolean is_adhoc_;
-    public long[] partition_bid;
+    public double[] partition_bid;
     //    public long lock_ratio;
     public long index_time;
     public long ts_allocation;
@@ -21,7 +21,7 @@ public class TxnContext {
     private int txn_type_;
     private SchemaRecordRef record_ref;
 
-    public TxnContext(int thread_Id, int fid, long bid) {
+    public TxnContext(int thread_Id, int fid, double bid) {
         this.thread_Id = thread_Id;
         this.thisOpId = null;
         this.fid = fid;
@@ -40,7 +40,7 @@ public class TxnContext {
      * @param fid
      * @param bid
      */
-    public TxnContext(String thisOpId, int fid, long bid) {
+    public TxnContext(String thisOpId, int fid, double bid) {
         this.thisOpId = thisOpId;
         this.fid = fid;
         this.bid = bid;
@@ -60,31 +60,31 @@ public class TxnContext {
      * @param bid
      * @param record_ref
      */
-    public TxnContext(String thisOpId, int fid, long bid, SchemaRecordRef record_ref) {
+    public TxnContext(String thisOpId, int fid, double bid, SchemaRecordRef record_ref) {
         this(thisOpId, fid, bid);
     }
 
-    public TxnContext(int thread_Id, int fid, long bid, SchemaRecordRef record_ref) {
+    public TxnContext(int thread_Id, int fid, double bid, SchemaRecordRef record_ref) {
         this(thread_Id, fid, bid);
         this.record_ref = record_ref;
     }
 
-    public TxnContext(String thisOpId, int fid, long bid, double[] index_time) {
+    public TxnContext(String thisOpId, int fid, double bid, double[] index_time) {
         this(thisOpId, fid, bid);
         index_time_ = index_time;
     }
 
-    public TxnContext(int thread_Id, int fid, long bid, double[] index_time) {
+    public TxnContext(int thread_Id, int fid, double bid, double[] index_time) {
         this(thread_Id, fid, bid);
         index_time_ = index_time;
     }
 
-    public TxnContext(int thread_id, int fid, long bid, int pid) {
+    public TxnContext(int thread_id, int fid, double bid, int pid) {
         this(thread_id, fid, bid);
         this.pid = pid;
     }
 
-    public TxnContext(int thread_id, int fid, long[] bid, int pid) {
+    public TxnContext(int thread_id, int fid, double[] bid, int pid) {
         this(thread_id, fid, bid[0]);
         this.partition_bid = bid;
         this.pid = pid;
@@ -98,7 +98,7 @@ public class TxnContext {
         return fid;
     }
 
-    public long getBID() {
+    public double getBID() {
         return bid;
     }
 }
