+++ state-engine/src/main/java/transaction/impl/ordered/TxnManagerTStream.java
@@ -55,7 +55,7 @@ public class TxnManagerTStream extends TxnManagerDedicatedAsy {
      * @return time spend in tp evaluation.
      */
     @Override
-    public void start_evaluate(int thread_Id, long mark_ID, int num_events) throws InterruptedException, BrokenBarrierException {
+    public void start_evaluate(int thread_Id, double mark_ID, int num_events) throws InterruptedException, BrokenBarrierException {
         MeasureTools.BEGIN_SCHEDULE_EXPLORE_TIME_MEASURE(thread_Id);
         SOURCE_CONTROL.getInstance().preStateAccessBarrier(thread_Id);//sync for all threads to come to this line to ensure chains are constructed for the current batch.
         if (enableGroup) {
